+++ js/main.js
@@ -1,6 +1,7 @@
 import { renderThumbnails } from './gallery-mini.js';
 import { setThumbnailsClick } from './gallery-fullsize.js';
-import { setFormSubmit, closeLoadForm, showSendSuccess } from './form.js';
+import { setFormSubmit, closeLoadForm } from './form.js';
+import { showSendSuccess } from './form-messages.js';
 import { getData } from './api.js';
 
 getData().then((data) => {
